[#1534] Rename the node_id files into unique_node_id
