+++ iceoryx2-cxx/include/iox2/iceoryx2.hpp
@@ -54,7 +54,6 @@
 #include "iox2/node.hpp"
 #include "iox2/node_details.hpp"
 #include "iox2/node_failure_enums.hpp"
-#include "iox2/node_id.hpp"
 #include "iox2/node_name.hpp"
 #include "iox2/node_state.hpp"
 #include "iox2/node_wait_failure.hpp"
@@ -120,6 +119,7 @@
 #include "iox2/type_name.hpp"
 #include "iox2/type_variant.hpp"
 #include "iox2/unable_to_deliver_strategy.hpp"
+#include "iox2/unique_node_id.hpp"
 #include "iox2/unique_port_id.hpp"
 #include "iox2/waitset.hpp"
 #include "iox2/waitset_enums.hpp"
