+++ iceoryx2-ffi/c/src/api/mod.rs
@@ -40,7 +40,6 @@ mod log;
 mod message_type_details;
 mod node;
 mod node_builder;
-mod node_id;
 mod node_name;
 mod notifier;
 mod notifier_details;
@@ -89,6 +88,7 @@ mod subscriber;
 mod subscriber_details;
 mod unique_client_id;
 mod unique_listener_id;
+mod unique_node_id;
 mod unique_notifier_id;
 mod unique_publisher_id;
 mod unique_reader_id;
@@ -122,7 +122,6 @@ pub use listener_details::*;
 pub use message_type_details::*;
 pub use node::*;
 pub use node_builder::*;
-pub use node_id::*;
 pub use node_name::*;
 pub use notifier::*;
 pub use notifier_details::*;
@@ -171,6 +170,7 @@ pub use subscriber::*;
 pub use subscriber_details::*;
 pub use unique_client_id::*;
 pub use unique_listener_id::*;
+pub use unique_node_id::*;
 pub use unique_notifier_id::*;
 pub use unique_publisher_id::*;
 pub use unique_reader_id::*;
