+++ iceoryx2-ffi/python/src/lib.rs
@@ -39,7 +39,6 @@ pub mod message_type_details;
 pub mod messaging_pattern;
 pub mod node;
 pub mod node_builder;
-pub mod node_id;
 pub mod node_name;
 pub mod node_state;
 pub mod notifier;
@@ -95,6 +94,7 @@ pub mod type_variant;
 pub mod unable_to_deliver_strategy;
 pub mod unique_client_id;
 pub mod unique_listener_id;
+pub mod unique_node_id;
 pub mod unique_notifier_id;
 pub mod unique_publisher_id;
 pub mod unique_reader_id;
@@ -149,7 +149,7 @@ fn _iceoryx2(py: Python<'_>, m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<crate::message_type_details::MessageTypeDetails>()?;
     m.add_class::<crate::node::Node>()?;
     m.add_class::<crate::node_builder::NodeBuilder>()?;
-    m.add_class::<crate::node_id::UniqueNodeId>()?;
+    m.add_class::<crate::unique_node_id::UniqueNodeId>()?;
     m.add_class::<crate::node_name::NodeName>()?;
     m.add_class::<crate::node_state::NodeState>()?;
     m.add_class::<crate::node_state::AliveNodeView>()?;
