+++ iceoryx2-ffi/python/src/port_factory_blackboard.rs
@@ -17,7 +17,6 @@ use pyo3::prelude::*;
 
 use crate::attribute_set::AttributeSet;
 use crate::error::NodeListFailure;
-use crate::node_id::UniqueNodeId;
 use crate::node_state::{
     AliveNodeView, AliveNodeViewType, DeadNodeView, DeadNodeViewType, NodeState,
 };
@@ -28,6 +27,7 @@ use crate::service_hash::ServiceHash;
 use crate::service_name::ServiceName;
 use crate::static_config_blackboard::StaticConfigBlackboard;
 use crate::type_storage::TypeStorage;
+use crate::unique_node_id::UniqueNodeId;
 
 pub(crate) enum PortFactoryBlackboardType {
     Ipc(
